Remove dependence on Param and add notes.
